Move s3 signing functions out of views.py
So I can use them directly from Mediathread without having similar code
in both places.
